+++ CHANGELOG.md
@@ -7,7 +7,6 @@ All notable changes to this project will be documented in this file.
 - Caching support
 - Artifact support
 - Custom build image support
-- Environment variable support
 - Support for non-GitHub sources
 
 ## [0.0.2] - 2023-07-26
@@ -15,6 +14,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Added support for boundary policy input.
+- Added environment variable support.
 
 ## [0.0.1] - 2023-07-20